requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-configurations==2.2
django-extensions==3.0.9
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
kombu==5.0.2
prompt-toolkit==3.0.8
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
psycopg2-binary==2.8.5
